requirements.txt
pip==24.1.2
Flask==2.0.1
instaloader==4.11
ffmpeg-python==0.2.0
vosk==0.3.30
requests==2.25.1
gunicorn==20.1.0
jsonpointer==2.1
setuptools==69.5.1
Werkzeug==2.0.3
websocket-client
google-generativeai==0.1.0  # Add the Google Generative AI SDK
gevent
python-dotenv